requirements.txt
discord.py>=2.3.0
icalendar
python-dotenv
requests
fastapi
uvicorn[standard]
gunicorn
motor>=3.1.1
dnspython>=2.1.0
slowapi>=0.1.5
email-validator>=1.1.3